Workable solution - still missing the comparsion of existing pregenerated different tests

+++ metrics_utility/test/snapshot_tests/snapshot_utils.py
@@ -13,10 +13,22 @@ import openpyxl
 import openpyxl.utils
 from openpyxl.worksheet.worksheet import Worksheet
 
+import warnings
+warnings.filterwarnings("ignore", category=ResourceWarning)
+
 @dataclass
 class DataShape:
     env_vars: Dict[str, str]
     params: List[str]
+    # values:
+    # run_command: 'Yes' - runs command using subprocess, 
+    # in future we may want to run function directly due to mocking datetime
+
+    # generated: date when report was generated
+    # run_manualy : 'Yes' - this will disable automatic testing between original and newly generated report
+    # it is used in situations when we want to manually compare existing original reports, for example
+    # reports with different params that should hold the same result 
+    # (for example month 2024-02 compared to 2024-02-01 to 2024-02-29)
     custom_params: Dict[str, str]
 
 def create_directory_if_not_exists(directory_path):
@@ -39,9 +51,8 @@ def parse_json_file(file_path):
     return None
 
 def save_snapshot_definition(data: DataShape, path: str):
-    pprint(path)
-
     try:
+        os.makedirs(os.path.dirname(path), exist_ok=True)
         # Open the file in write mode
         with open(path, 'w') as json_file:
             # Serialize the dictionary into JSON and write to the file
@@ -58,11 +69,16 @@ def find_json_files(directory):
     return list(Path(directory).rglob("*.json"))
 
 def run_and_generate_snapshot_definitions(directory):
+    print(f'\nRun and generate snapshots from {directory}\n')
     json_files = find_json_files(directory)
+
     for json_file in json_files:
+        print(f'Found {json_file}')
         output_dir = json_file.as_posix().removesuffix('.json')
         create_directory_if_not_exists(output_dir)
-        generated_file = run_snapshot_definition(json_file)
+        
+        data = parse_json_file(json_file)
+        generated_file = run_snapshot_definition(data)
         output_file = output_dir + "/report.xlsx"
 
         if os.path.exists(output_file):
@@ -72,6 +88,49 @@ def run_and_generate_snapshot_definitions(directory):
         print(f'Report generated and moved from {generated_file} to {output_file}\n')
     return
 
+# Returns path to generated file
+def run_snapshot_definition(data):
+    env_vars = copy.deepcopy(data['env_vars'])
+    env_vars["AWX_LOGGING_MODE"] = 'stdout'
+    python_executable = sys.executable
+    
+    params = []
+    params.append(python_executable)
+    params.extend(data['params'])
+
+    # Runs command, in future, we want to support also calling the test function directly due to mocking datetime.now
+    if (data['custom_params']['run_command'] == 'Yes'):
+        result = subprocess.run(
+            params,
+            text=True,
+            stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE,
+            env=env_vars,
+        )
+
+        if result.returncode != 0:
+            print('Generating of report failed')
+            print(result.stderr)
+            return
+        
+        text = result.stderr + '/n' + result.stdout
+    else:
+        return ''
+    
+    generated_file = get_file_name(params, env_vars)
+
+    if generated_file:
+        return generated_file
+        
+    # Regular expression to capture the file path if get_file_name was not able to compute it
+    pattern = r"Report generated into directory:\s*(.*?)\.xlsx"
+
+    match = re.search(pattern, text)
+    if match:
+        return match.group(1) + '.xlsx'
+
+    return ''
+
 def run_and_test_snapshot_definitions(directory):
     json_files = find_json_files(directory)
     print('Found definitions:')
@@ -79,17 +138,44 @@ def run_and_test_snapshot_definitions(directory):
     for json_file in json_files:
         print('')
         data_dir = json_file.as_posix().removesuffix('.json')
-        generated_file = run_snapshot_definition(json_file)
+        data : DataShape = parse_json_file(json_file)
+
         original_file = './' + data_dir + "/report.xlsx"
-
-        # compare the generated and original_file
-        compare_CCSPv2_reports(original_file, generated_file)
-
-        if os.path.exists(generated_file):
-            print(f'Removing {generated_file}')
-            os.remove(generated_file)
         
-        return
+        run_manualy = None
+        if 'run_manualy' in data['custom_params']:
+            run_manualy = data['custom_params']
+        if (run_manualy != 'Yes'):
+            generated_file = run_snapshot_definition(data)
+
+            # compare the generated and original_file
+            if (data["env_vars"]["METRICS_UTILITY_REPORT_TYPE"] == 'CCSPv2'):
+                print(f'Compare {original_file} to {generated_file}')
+                compare_CCSPv2_reports(original_file, generated_file)
+
+            if os.path.exists(generated_file):
+                print(f'Removing {generated_file}')
+                os.remove(generated_file)
+        else:
+            print(f'Skipping the file {original_file} because it has custom_params run_manualy to Yes')
+        
+    return
+
+def compare_CCSPv2_reports(original_report_path, generated_report_path):
+    print(f'Opening {generated_report_path}')
+    g_wb = openpyxl.load_workbook(filename = generated_report_path)  
+                      
+    print(f'Opening {original_report_path}')
+    o_wb = openpyxl.load_workbook(filename = original_report_path)
+
+    try:
+        compareWorksheets(g_wb, o_wb, 0, ['H1', 'B5'])
+        compareWorksheets(g_wb, o_wb, 1, [])
+        compareWorksheets(g_wb, o_wb, 2, [])
+    finally:
+        g_wb.close()
+        o_wb.close()
+       
     return
 
 def compareWorksheets(workbook_generated, workbook_original, sheetNumber, exceptions : List[str]):   
@@ -123,74 +209,12 @@ def compareWorksheets(workbook_generated, workbook_original, sheetNumber, except
                 assert (
                     valG == valO
                     ), f"Column names do not match for sheet number: {sheetNumber}. Address {addr}. Actual value = {valG}, expected value = {valO}"
-           
-        
-
-def compare_CCSPv2_reports(original_report, generated_report):
-    print(f'Opening {generated_report}')
-   
-    g_wb = openpyxl.load_workbook(filename = generated_report,  
-                          read_only=True)
-    
-    print(f'Opening {original_report}')
-    o_wb = openpyxl.load_workbook(filename = original_report, 
-                          read_only=True)
-
     
 
-    compareWorksheets(g_wb, o_wb, 0, ['H1'])
-    compareWorksheets(g_wb, o_wb, 1, [])
-    compareWorksheets(g_wb, o_wb, 2, [])
-
-    return
 
 def get_lines_count():
     return
 
-# Returns path to generated file
-def run_snapshot_definition(file_path):
-    data : DataShape = parse_json_file(file_path)
-    
-    env_vars = copy.deepcopy(data['env_vars'])
-    env_vars["AWX_LOGGING_MODE"] = 'stdout'
-    python_executable = sys.executable
-    
-    params = []
-    params.append(python_executable)
-    params.extend(data['params'])
-
-    # Runs command, in future, we want to support also calling the test function directly due to mocking datetime.now
-    if (data['custom_params']['run_command'] == True):
-        result = subprocess.run(
-            params,
-            text=True,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            env=env_vars,
-        )
-
-        if result.returncode != 0:
-            print('Generating of report failed')
-            print(result.stderr)
-            return
-        
-        text = result.stderr + '/n' + result.stdout
-    else:
-        return ''
-    
-    generated_file = get_file_name(params, env_vars)
-
-    if generated_file:
-        return generated_file
-        
-    # Regular expression to capture the file path if get_file_name was not able to compute it
-    pattern = r"Report generated into directory:\s*(.*?)\.xlsx"
-
-    match = re.search(pattern, text)
-    if match:
-        return match.group(1) + '.xlsx'
-
-    return ''
 
 def get_file_name(params, env_vars):
     if env_vars['METRICS_UTILITY_REPORT_TYPE'] == 'CCSPv2':
